Quiz_master_23f2004341/requirements.txt
Flask
Flask-Login
Flask-WTF
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Caching
Flask-JWT-Extended
redis
celery
python-dotenv
requests
